refactor(map-calibration): promote contracts-tier types out of Detection/ and .Capture (Phase B)
Types promoted to Mithril.MapCalibration root (the contracts tier):
- Domain types: MapRect, LandmarkReference, CandidateTransform, CanonicalLandmarkTypes
- Boundary interfaces: ICalibrationDetector, ICalibrationConfidenceGate,
  IBaseTextureProvider, IIconTemplateProvider, IAssetExtractor
- Refiner contracts: IMapRegionRefiner, MapRegionRefineResult, LocateMetrics

Service-tier relocation:
- ProcessAssetExtractor moves from Detection/ to Mithril.MapCalibration/Internal/

Spec: docs/planning/map-calibration-detection-project-split/spec.md
