+++ src/Mithril.MapCalibration/IAssetExtractor.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Mithril.MapCalibration.Detection;
+namespace Mithril.MapCalibration;
 
 /// <summary>
 /// The single <c>src/**</c> touchpoint that knows the out-of-process
